% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chat1.R
\name{Chat1}
\alias{Chat1}
\title{\code{Chat1} simple CQ model}
\usage{
Chat1(params, flow)
}
\value{
A vector of concentration predictions for each time-step
}
\description{
This is a model function. It predicts in-stream concentration.
}
\keyword{model}
