% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getParam}
\alias{getParam}
\title{Get Param}
\usage{
getParam(
  Chat.model.names = c("Chat1", "Chat13"),
  input.data = data_all,
  cmaes.results = model,
  model.setup = model.setup
)
}
\arguments{
\item{Chat.model.names}{character string vector with a 'Chat#' model name from provided models (i.e. Chat1-Chat15). Chat1 and Chat13 default}

\item{input.data}{dataframe of daily runoff and concentration AND predictions of concentration and baseflow}

\item{cmaes.results}{list of cmaes.results from fitted models}

\item{model.setup}{lists of details about data, model, and site from setModels()}
}
\value{
output summary dataframe with statistics (negLL, AIC, NSE, RMSE, and BFI)
}
\description{
getParam outputs the parameters for each fitted C-Q model
}
\details{
\code{getParam}


exported datafame with parameters from fitted C-Q models
}
\keyword{getParam}
