% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getStats}
\alias{getStats}
\title{Get Stats}
\usage{
getStats(
  Chat.model.names = c("Chat1", "Chat13"),
  input.data = data_all,
  cmaes.results = models,
  model.setup = model.setup
)
}
\arguments{
\item{Chat.model.names}{character string vector with a 'Chat#' model name from provided models (i.e. Chat1-Chat15). Chat1 and Chat13 default}

\item{input.data}{dataframe of daily runoff and concentration AND predictions of concentration and baseflow}

\item{cmaes.results}{list of cmaes.results from fitted models}

\item{model.setup}{lists of details about data, model, and site from setModels()}
}
\value{
output summary dataframe with statistics (negLL, AIC, NSE, RMSE, and BFI)
}
\description{
getStats calculates the statistics for each fitted C-Q model
}
\details{
\code{getStats}


exported summary table with negLL, AIC, NSE, RMSE, and BFI
}
\keyword{getStats}
