% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{plotResults}
\alias{plotResults}
\title{plot Results}
\usage{
plotResults(
  Chat.model.names = c("Chat1", "Chat13"),
  input.data = data.frame(),
  model.setup = model.setup
)
}
\arguments{
\item{Chat.model.names}{character string vector with a 'Chat#' model name from provided models (i.e. Chat1-Chat15). Chat1 and Chat13 default}

\item{input.data}{dataframe of daily runoff and concentration AND predictions of concentration and baseflow}

\item{model.setup}{lists of details about data, model, and site from setModels()}
}
\value{
annual timeseries plot comparing predictions from two models with observations, streamflow and baseflow; C-Q scatter plots of each model
}
\description{
plotResults plots predictions from each fitted C-Q model
}
\details{
\code{plotResults}


plots with parameters from fitted C-Q models
}
\keyword{plotResults}
