% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{runModels}
\alias{runModels}
\title{Run models}
\usage{
runModels(model.setup = list())
}
\arguments{
\item{model.setup}{lists of details about data, model, and site from setModels()}
}
\value{
fitted C-Q models
}
\description{
Run C-Q models for comparison
}
\details{
\code{runModels}


Runs C-Q models from the selection of Chat1-15. Default runs simple C-Q model (Chat1) and quick-slow Hubbard Brook model (Chat13)
}
\keyword{runModels}
