% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{setModels}
\alias{setModels}
\title{Set models}
\usage{
setModels(
  Chat.model.names = c("Chat1", "Chat13"),
  input.data = data_all,
  Qthresh = 0,
  Likelihood.name = "GaussLiklihood",
  site.id = "",
  site.name = ""
)
}
\arguments{
\item{Chat.model.names}{character string vector with a 'Chat#' model name from provided models (i.e. Chat1-Chat15). Chat1 and Chat13 default}

\item{input.data}{dataframe of daily runoff and concentration. colnames = c("year", "month", "day", "C", "flow_mm_d")}

\item{Qthresh}{numeric low-flow streamflow threshold, models only fitted to observations with same day streamflow above this threshold.}

\item{Likelihood.name}{character string with name of likelihood function ("GaussLiklihood","GaussLiklihoodAR1", or "GaussLiklihoodAR3")}

\item{site.id}{character string with identifier of gauge or catchment}

\item{site.name}{character string with name of gauge or catchment}
}
\value{
C-Q models ready for runModels
}
\description{
Set-up C-Q models for comparison
}
\details{
\code{setModels}


Sets-up C-Q models from the selection of Chat1-15. Default sets-up simple C-Q model (Chat1) and quick-slow Hubbard Brook model (Chat13)
}
\keyword{setModels}
