% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getParam}
\alias{getParam}
\title{Get Param}
\usage{
getParam(
  model.setup = list(),
  Chat.model.name = character(),
  output.data = dataframe()
)
}
\arguments{
\item{model.setup}{lists of details about data, model, and site from \code{setModels}}

\item{output.data}{dataframe of daily concentration and baseflow predictions from \code{getResults}}
}
\value{
output summary dataframe with parameters for fitted models
}
\description{
getParam outputs the parameters for each fitted C-Q model
}
\details{
\code{getParam}


exported datafame with parameters from fitted C-Q models
}
\keyword{getParam}
