% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getResults}
\alias{getResults}
\title{Get Results}
\usage{
getResults(model.setup = list())
}
\arguments{
\item{model.setup}{lists of details about data, model, and site from setModels()}

\item{cmaes.results}{list of cmaes.results from fitted models}
}
\value{
output original data_all dataframe with predicted concentration and baseflow of fitted models
}
\description{
getResults of fitted C-Q models
}
\details{
\code{getResults}


exported predicted concentration and baseflow of fitted models
}
\keyword{getResults}
