% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getStats}
\alias{getStats}
\title{Get Stats}
\usage{
getStats(model.setup = list(), output.data = dataframe())
}
\arguments{
\item{model.setup}{lists of details about data, model, and site from \code{setModels}}

\item{output.data}{dataframe of daily concentration and baseflow predictions from \code{getResults}}
}
\value{
output summary dataframe with statistics (negLL, AIC, NSE, RMSE, and BFI)
}
\description{
getStats calculates the statistics for each fitted C-Q model
}
\details{
\code{getStats}


exported summary table with negLL, AIC, NSE, RMSE, and BFI
}
\keyword{getStats}
